backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
openai
sqlalchemy
pymysql
python-dotenv